feat: #661 Remove old trigger entrypoint, tidy
